STAR-2408 Hack around libyaml 0.2.5 behavior diffs
[STAR-2408](https://outreach-io.atlassian.net/browse/STAR-2408)

The folks behind libyaml decided to change it so it was less likely to
emit trailing whitespace.  That's good for them, but bad for us.  We've
been using MD5 hashes of those YAML dumps for our recordings and now
those MD5 hash values are different.

Eventually we'll want to settle towards the new behavior.  But we can't
just go updating the hashes; some dev machines / build bots / whatever
still have the old libyaml and would be confused by the new hashes.

This ugly hack is good enough to keep tests passing in either libyaml
version, for now.  Hopefully it'll get us over the hump and we can
remove it and re-record the tests alter.

I think this is the PR that changed the behavior:
https://github.com/yaml/libyaml/issues/46